app/pyproject.toml
[project]
name = "bagel-store"
version = "1.0.0"
description = "Bagel Store Ordering Application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
]
